+++ tests/Repository/ProducersRepositoryTest.php
@@ -24,14 +24,13 @@ class ProducersRepositoryTest extends PHPUnit_Framework_TestCase
     public function setUp()
     {
         $oauth2 = "1/token111accessoauth2";
-        $cacheMock = $this->getMock('Nokaut\ApiKit\Cache\CacheInterface');
-        $loggerMock = $this->getMock('Psr\Log\LoggerInterface');
+        $cacheMock = $this->createMock('Nokaut\ApiKit\Cache\CacheInterface');
+        $loggerMock = $this->createMock('Psr\Log\LoggerInterface');
         $client = $this->getMockBuilder('\GuzzleHttp\Client')->disableOriginalConstructor()->getMock();
-        $this->clientApiMock = $this->getMock(
-            'Nokaut\ApiKit\ClientApi\Rest\RestClientApi',
-            array('convertResponse', 'getClient', 'log', 'convertResponseToSaveCache'),
-            array($loggerMock, $oauth2)
-        );
+        $this->clientApiMock = $this->getMockBuilder('Nokaut\ApiKit\ClientApi\Rest\RestClientApi')
+            ->setMethods(['convertResponse', 'getClient', 'log', 'convertResponseToSaveCache'])
+            ->setConstructorArgs([$loggerMock, $oauth2])
+            ->getMock();
         $this->clientApiMock->expects($this->any())->method('getClient')
             ->will($this->returnValue($client));